proj-1/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
protobuf==3.19.1
six==1.16.0
